Several improvements.
- Revamped how assemblies are loaded to achieve better
parallelism at lower memory costs. This speeds things
up when dealing with lots of inputs.

- Made the default tool output 'just right' to keep users
informed of on-going progress.

- Added the --quiet option to prevent all informational
output.

- Made early duplicate detection case-insensitive which finds a
lot more cases in the real world.

- Removed version information from the duplicate assembly
report, since we generally don't have this info anymore.

- Improve the graph dump format to make it easier to search
for things.

- Fix the hacked-up system types that are explicitly created
when the core library is not supplied for analysis. There were
a few mistyped symbols (wrong casing) and the hack wasn't applied in every case it is needed. These fixes should remove
swaths of false positives.